requirements.txt
## The following requirements were added by pip freeze:
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.4
filelock==3.13.3
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
gensim==4.3.2
greenlet==3.0.3
huggingface-hub==0.22.2
idna==3.6
importlib_resources==6.4.0
jieba==0.42.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain-core==0.1.39
langsmith==0.1.39
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
orjson==3.10.0
packaging==23.2
pandas==2.2.1
pillow==10.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
pytz==2024.1
pyvi==0.1.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scikit-surprise==1.1.3
scipy==1.13.0
seaborn==0.13.2
sentence-transformers==2.6.1
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==7.0.4
SQLAlchemy==2.0.29
surprise==0.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
underthesea==6.8.0
underthesea_core==1.0.4
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
